✨ Split of Neutroics pre cell csg and openmc running (#3243)
* 🎨 Neutronics PreCell / openmc separation

* 🚚 Move stuff about

* 🎨 Reduce surface that requires openmc

* 🚚 Move materials about

* 🎨 Initial material rearrangement

* 🌐 More internationalisation

* ♻️ Neutronics material movement

* 🎨 More materials stuff

* ♻️ Cleanup

* 🎨 Materials cleanup
