+++ bluemira/neutronics/params.py
@@ -8,106 +8,54 @@
 from __future__ import annotations
 
 from dataclasses import dataclass, fields
-from enum import Enum, auto
-from pathlib import Path
-from typing import TYPE_CHECKING, Literal
 
 import numpy as np
-import openmc
 
 from bluemira.base.constants import raw_uc
 from bluemira.base.look_and_feel import bluemira_warn
+from bluemira.base.parameter_frame import Parameter, ParameterFrame
 from bluemira.geometry.error import GeometryError
-
-if TYPE_CHECKING:
-    from pathlib import Path
-
-    from bluemira.base.parameter_frame import ParameterFrame
-
-
-class BlanketLayers(Enum):
-    """
-    The five layers of the blanket as used in the neutronics simulation.
-
-    Variables
-    ---------
-    Surface
-        The surface layer of the first wall.
-    First wall
-        Typically made of tungsten or Eurofer
-    BreedingZone
-        Where tritium is bred
-    Manifold
-        The pipe works and supporting structure
-    VacuumVessel
-        The vacuum vessel keeping the plasma from mixing with outside air.
-    """
-
-    Surface = auto()
-    FirstWall = auto()
-    BreedingZone = auto()
-    Manifold = auto()
-    VacuumVessel = auto()
 
 
 @dataclass
-class OpenMCSimulationRuntimeParameters:
-    """Parameters used in the actual simulation
+class OpenMCNeutronicsSolverParams(ParameterFrame):
+    """
 
     Parameters
     ----------
-    particles:
-        Number of neutrons emitted by the plasma source per batch.
-    batches:
-        How many batches to simulate.
-    photon_transport:
-        Whether to simulate the transport of photons (i.e. gamma-rays created) or not.
-    electron_treatment:
-        The way in which OpenMC handles secondary charged particles.
-        'thick-target bremsstrahlung' or 'local energy deposition'
-        'thick-target bremsstrahlung' accounts for the energy carried away by
-        bremsstrahlung photons and deposited elsewhere, whereas 'local energy
-        deposition' assumes electrons deposit all energies locally.
-        (the latter is expected to be computationally faster.)
-    run_mode:
-        see below for details:
-        https://docs.openmc.org/en/stable/usersguide/settings.html#run-modes
-    openmc_write_summary:
-        whether openmc should write a 'summary.h5' file or not.
-    cross_section_xml:
-        Where the xml file for cross-section is stored locally.
+    major_radius:
+        Major radius of the machine
+    aspect_ratio:
+        aspect ratio of the machine
+    elongation:
+        elongation of the plasma
+    triangularity:
+        triangularity of the plasma
+    reactor_power:
+        total reactor (thermal) power when operating at 100%
+    peaking_factor:
+        (max. heat flux on fw)/(avg. heat flux on fw)
+    temperature:
+        plasma temperature (assumed to be uniform throughout the plasma)
+    shaf_shift:
+        Shafranov shift
+        shift of the centre of flux surfaces, i.e.
+        mean(min radius, max radius) of the LCFS,
+        towards the outboard radial direction.
+    vertical_shift:
+        how far (upwards) in the z direction is the centre of the plasma
+        shifted compared to the geometric center of the poloidal cross-section.
     """
 
-    # Parameters used outside of setup_openmc()
-    particles: int  # number of particles used in the neutronics simulation
-    cross_section_xml: str | Path
-    batches: int = 2
-    photon_transport: bool = True
-    # Bremsstrahlung only matters for very thin objects
-    electron_treatment: Literal["ttb", "led"] = "led"
-    run_mode: str = openmc.settings.RunMode.FIXED_SOURCE.value
-    openmc_write_summary: bool = False
-    parametric_source: bool = True
-    plot_axis: str = "xz"
-    plot_pixel_per_metre: int = 100
-
-
-class BlanketType(Enum):
-    """Types of allowed blankets, named by their acronyms."""
-
-    DCLL = auto()
-    HCPB = auto()
-    WCLL = auto()
-
-
-@dataclass
-class BreederTypeParameters:
-    """Dataclass to hold information about the breeder blanket material
-    and design choices.
-    """
-
-    enrichment_fraction_Li6: float
-    blanket_type: BlanketType
+    major_radius: Parameter[float]  # [m]
+    aspect_ratio: Parameter[float]  # [dimensionless]
+    elongation: Parameter[float]  # [dimensionless]
+    triangularity: Parameter[float]  # [dimensionless]
+    reactor_power: Parameter[float]  # [W]
+    peaking_factor: Parameter[float]  # [dimensionless]
+    temperature: Parameter[float]  # [K]
+    shaf_shift: Parameter[float]  # [m]
+    vertical_shift: Parameter[float]  # [m]
 
 
 @dataclass(frozen=True)
@@ -300,151 +248,27 @@ class TokamakDimensions:
     central_solenoid: ToroidalFieldCoilDimension
 
     @classmethod
-    def from_tokamak_geometry(
+    def from_parameterframe(
         cls,
-        tokamak_geometry: TokamakGeometry,
-        blanket_io_cut: float,
-        tf_inner_radius: float,
-        tf_outer_radius: float,
-        divertor_surface_tk: float = 0.1,
-        blanket_surface_tk: float = 0.01,
-        blk_ib_manifold: float = 0.02,
-        blk_ob_manifold: float = 0.2,
+        params,
     ):
-        """Bodge method that can be deleted later once
-        :func:`~get_preset_physical_properties` migrated over to use TokamakDimensions.
-        """
+        """Setup tokamak dimensions"""
         return cls(
             BlanketThickness(
-                blanket_surface_tk,
-                tokamak_geometry.inb_fw_thick,
-                tokamak_geometry.inb_bz_thick,
-                blk_ib_manifold,
+                params.blanket_surface_tk.value,
+                params.inboard_fw_tk.value,
+                params.inboard_breeding_tk.value,
+                params.blk_ib_manifold.value,
             ),
-            blanket_io_cut,
+            params.blanket_io_cut.value,
             BlanketThickness(
-                blanket_surface_tk,
-                tokamak_geometry.outb_fw_thick,
-                tokamak_geometry.outb_bz_thick,
-                blk_ob_manifold,
+                params.blanket_surface_tk.value,
+                params.outboard_fw_tk.value,
+                params.outboard_breeding_tk.value,
+                params.blk_ob_manifold.value,
             ),
-            DivertorThickness(divertor_surface_tk),
-            ToroidalFieldCoilDimension(tf_inner_radius, tf_outer_radius),
+            DivertorThickness(params.divertor_surface_tk.value),
+            ToroidalFieldCoilDimension(
+                params.tf_inner_radius.value, params.tf_outer_radius.value
+            ),
         )
-
-
-@dataclass(frozen=True)
-class TokamakGeometry:
-    """The thickness measurements for all of the generic components of the tokamak.
-
-    Parameters
-    ----------
-    inb_fw_thick:
-        inboard first wall thickness [m]
-    inb_bz_thick:
-        inboard breeding zone thickness [m]
-    inb_mnfld_thick:
-        inboard manifold thickness [m]
-    inb_vv_thick:
-        inboard vacuum vessel thickness [m]
-    tf_thick:
-        toroidal field coil thickness [m]
-    outb_fw_thick:
-        outboard first wall thickness [m]
-    outb_bz_thick:
-        outboard breeding zone thickness [m]
-    outb_mnfld_thick:
-        outboard manifold thickness [m]
-    outb_vv_thick:
-        outboard vacuum vessel thickness [m]
-    inb_gap:
-        inboard gap [m]
-    """
-
-    inb_fw_thick: float
-    inb_bz_thick: float
-    inb_mnfld_thick: float
-    inb_vv_thick: float
-    tf_thick: float
-    outb_fw_thick: float
-    outb_bz_thick: float
-    outb_mnfld_thick: float
-    outb_vv_thick: float
-    inb_gap: float
-
-
-def get_preset_physical_properties(
-    blanket_type: str | BlanketType,
-) -> tuple[BreederTypeParameters, TokamakGeometry]:
-    """
-    Works as a switch-case for choosing the tokamak geometry
-    and blankets for a given blanket type.
-    The allowed list of blanket types are specified in BlanketType.
-    Currently, the blanket types with pre-populated data in this function are:
-    {'wcll', 'dcll', 'hcpb'}
-    """
-    if not isinstance(blanket_type, BlanketType):
-        blanket_type = BlanketType[blanket_type.lower()]
-
-    breeder_materials = BreederTypeParameters(
-        blanket_type=blanket_type,
-        enrichment_fraction_Li6=0.60,
-    )
-
-    # Geometry variables
-
-    # Break down from here.
-    # Paper inboard build ---
-    # Nuclear analyses of solid breeder blanket options for DEMO:
-    # Status,challenges and outlook,
-    # Pereslavtsev, 2019
-    #
-    # 0.400,      # TF Coil inner
-    # 0.200,      # gap                  from figures
-    # 0.060,       # VV steel wall
-    # 0.480,      # VV
-    # 0.060,       # VV steel wall
-    # 0.020,       # gap                  from figures
-    # 0.350,      # Back Supporting Structure
-    # 0.060,       # Back Wall and Gas Collectors   Back wall = 3.0
-    # 0.350,      # breeder zone
-    # 0.022        # fw and armour
-
-    shared_building_geometry = {  # that are identical in all three types of reactors.
-        "inb_gap": 0.2,  # [m]
-        "inb_vv_thick": 0.6,  # [m]
-        "tf_thick": 0.4,  # [m]
-        "outb_vv_thick": 0.6,  # [m]
-    }
-    if blanket_type is BlanketType.WCLL:
-        tokamak_geometry = TokamakGeometry(
-            **shared_building_geometry,
-            inb_fw_thick=0.027,  # [m]
-            inb_bz_thick=0.378,  # [m]
-            inb_mnfld_thick=0.435,  # [m]
-            outb_fw_thick=0.027,  # [m]
-            outb_bz_thick=0.538,  # [m]
-            outb_mnfld_thick=0.429,  # [m]
-        )
-    elif blanket_type is BlanketType.DCLL:
-        tokamak_geometry = TokamakGeometry(
-            **shared_building_geometry,
-            inb_fw_thick=0.022,  # [m]
-            inb_bz_thick=0.300,  # [m]
-            inb_mnfld_thick=0.178,  # [m]
-            outb_fw_thick=0.022,  # [m]
-            outb_bz_thick=0.640,  # [m]
-            outb_mnfld_thick=0.248,  # [m]
-        )
-    elif blanket_type is BlanketType.HCPB:
-        # HCPB Design Report, 26/07/2019
-        tokamak_geometry = TokamakGeometry(
-            **shared_building_geometry,
-            inb_fw_thick=0.027,  # [m]
-            inb_bz_thick=0.460,  # [m]
-            inb_mnfld_thick=0.560,  # [m]
-            outb_fw_thick=0.027,  # [m]
-            outb_bz_thick=0.460,  # [m]
-            outb_mnfld_thick=0.560,  # [m]
-        )
-    return breeder_materials, tokamak_geometry
